refactor(admin-config): Rename "endpoints" package to "routes"
Align on using the Ktor terminology to not have two different names for
the same thing. The same change will be done to the plugin-manager
component later.
